mopping up refactor changes
